+++ src/components/buttons/PlaceholderButton/PlaceholderButton.tsx
@@ -4,7 +4,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { IconName } from '@fortawesome/fontawesome-svg-core'
 import { jsx } from '@emotion/core'
 
-import { processButtonContainer } from '../process-button-container'
+import { processButtonContainer } from '../processButtonContainer'
 
 import { addButtonStyles, disabledStyles } from './styles'
